refactor: move handlers into submodules
