+++ crates/http/src/handlers/mod.rs
@@ -1,0 +1,3 @@
+pub mod auth;
+pub mod pages;
+pub mod user;
